requirements.txt
langgraph==0.6.6
langgraph-cli==0.4.0
langgraph-sdk==0.2.4
langchain==0.2.16
langchain-community==0.2.12
nats-py==2.6.0
redis==5.0.7
fastapi==0.111.0
uvicorn==0.30.1
prometheus_client==0.20.0
requests==2.32.3
aiohttp==3.9.5
pydantic==2.8.2
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
numpy==1.26.4
scipy==1.11.4
influxdb-client==1.42.0
orjson==3.10.7
